move user_info
