+++ ydb/core/persqueue/pqtablet/partition/ya.make
@@ -13,6 +13,7 @@ SRCS(
     partition_sourcemanager.cpp
     partition_write.cpp
     sourceid.cpp
+    user_info.cpp
 )
 
 
@@ -33,8 +34,4 @@ PEERDIR(
 END()
 
 RECURSE_FOR_TESTS(
-#    ut
-#    dread_cache_service/ut
-#    ut/slow
-#    ut/ut_with_sdk
 )
