+++ ydb/core/persqueue/public/constants.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <util/generic/string.h>
+
+namespace NKikimr::NPQ {
+
+static const TString CLIENTID_COMPACTION_CONSUMER = "__ydb_compaction_consumer";
+static const TString CLIENTID_WITHOUT_CONSUMER = "$without_consumer";
+
+}
