+++ ydb/core/persqueue/ut/common/pq_ut_common.h
@@ -2,7 +2,7 @@
 
 #include <ydb/core/persqueue/pq.h>
 #include <ydb/core/persqueue/events/internal.h>
-#include <ydb/core/persqueue/user_info.h>
+#include <ydb/core/persqueue/pqtablet/partition/user_info.h>
 #include <ydb/core/testlib/actors/test_runtime.h>
 #include <ydb/core/testlib/basics/runtime.h>
 #include <ydb/core/testlib/tablet_helpers.h>
