+++ ydb/services/lib/actors/pq_schema_actor.cpp
@@ -3,7 +3,7 @@
 #include <ydb/public/sdk/cpp/src/library/persqueue/obfuscate/obfuscate.h>
 #include <ydb/library/persqueue/topic_parser/topic_parser.h>
 #include <ydb/core/base/feature_flags.h>
-#include <ydb/core/persqueue/public/utils.h>
+#include <ydb/core/persqueue/public/constants.h>
 
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/library/jwt/jwt.h>
